Rename USBDevice folder to usb
