+++ BREAKING_CHANGES.md
@@ -1,0 +1,3 @@
+v1.1.4:
+
+* `USBDevice` folder renamed to `usb`
