+++ src/wch-ch56x-lib/usb/usb2_utils.h
@@ -30,8 +30,8 @@ limitations under the License.
 #pragma GCC diagnostic pop
 
 #include "wch-ch56x-lib/logging/logging.h"
-#include "wch-ch56x-lib/USBDevice/usb_device.h"
-#include "wch-ch56x-lib/USBDevice/usb_types.h"
+#include "wch-ch56x-lib/usb/usb_device.h"
+#include "wch-ch56x-lib/usb/usb_types.h"
 #include <stdbool.h>
 
 #ifdef __cplusplus
