+++ src/wch-ch56x-lib/usb/usb_endpoints.c
@@ -15,5 +15,5 @@ limitations under the License.
 
 *******************************************************************************/
 
-#include "wch-ch56x-lib/USBDevice/usb_endpoints.h"
-#include "wch-ch56x-lib/USBDevice/usb_device.h"
+#include "wch-ch56x-lib/usb/usb_endpoints.h"
+#include "wch-ch56x-lib/usb/usb_device.h"
