+++ tests/test_firmware_usb_loopback/User/usb2_device_descriptors.h
@@ -19,7 +19,7 @@ limitations under the License.
 #define USB2_DEVICE_DESCRIPTOR_H
 
 #include "definitions.h"
-#include "wch-ch56x-lib/USBDevice/usb_descriptors.h"
+#include "wch-ch56x-lib/usb/usb_descriptors.h"
 
 const uint8_t* usb2_device_configs[1];
