+++ tests/test_firmware_usb_loopback/User/usb_device.h
@@ -19,9 +19,9 @@ limitations under the License.
 #define USB_DEVICE_USER_H
 
 #include "definitions.h"
-#include "wch-ch56x-lib/USBDevice/usb_descriptors.h"
-#include "wch-ch56x-lib/USBDevice/usb_device.h"
-#include "wch-ch56x-lib/USBDevice/usb_endpoints.h"
+#include "wch-ch56x-lib/usb/usb_descriptors.h"
+#include "wch-ch56x-lib/usb/usb_device.h"
+#include "wch-ch56x-lib/usb/usb_endpoints.h"
 
 uint8_t hydradancer_product_string_descriptor[] = {
 	'H',
