+++ tools/firmware_debug_board/User/usb2_device_descriptors.h
@@ -18,7 +18,7 @@ limitations under the License.
 #ifndef USB2_DEVICE_DESCRIPTOR_H
 #define USB2_DEVICE_DESCRIPTOR_H
 
-#include "wch-ch56x-lib/USBDevice/usb_descriptors.h"
+#include "wch-ch56x-lib/usb/usb_descriptors.h"
 
 const uint8_t* usb2_device_configs[1];
